microservices/chatbot-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
cassandra-driver==3.28.0
requests==2.31.0
pydantic==2.5.3
